Missing benchmark is returned; Karma supports now unit tests and performance tests;
