+++ package.json
@@ -10,6 +10,7 @@
     "babel-plugin-transform-es2015-modules-umd": "^6.8.0",
     "babel-preset-es2015": "^6.6.0",
     "babel-preset-es2015-rollup": "^1.1.1",
+    "benchmark": "^2.1.0",
     "camelcase": "^3.0.0",
     "css-loader": "^0.23.1",
     "cz-conventional-changelog": "^1.1.6",
